delete more pages
