+++ src/pages/GiveParty/Bullseye.tsx
@@ -17,8 +17,6 @@ import {
   Tooltip,
 } from 'ui';
 
-import { abbreviateNumber } from './PartyStats';
-
 export interface User {
   username: string;
   avatar: string;
@@ -332,3 +330,17 @@ export const Bullseye = ({
     </Popover>
   );
 };
+
+export function abbreviateNumber(num: number): string {
+  if (num < 1000) {
+    return num.toString();
+  } else if (num < 10000) {
+    // Convert to thousands with one decimal place
+    let abbreviated = (num / 1000).toFixed(1);
+    abbreviated = abbreviated.replace(/\.0$/, '');
+    return abbreviated + 'k';
+  } else {
+    // For 10000 and above, round down to the nearest thousand
+    return Math.floor(num / 1000) + 'k';
+  }
+}
